Added the dns-options tool here.
That tool existed in the snapwebsites environment under the snapmanager.
It can be used to tweak the settings of BIND9 because you are forced to
do an edit (or delete) of the existing files (i.e. BIND9 does not offer
a way to add options _at a later time_; you have to change them in the
named.conf.options of similar file and that's it...)

This tool is a command line that allows you to script such modifications.