Next step toward a running version.
Added support for a CNAME.
Properly check IP addresses to generate A or AAAA fields (instead of just A)
Verify all the IPs (with libaddr) and domains (with libtld).
Added call to `named-checkzone` to verify that our zones are valid.
Moved the generation of the zone to the zone_files class.